third_party/WebKit/LayoutTests/fast/table/internal-center-ua-only.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style>
th {
  text-align: right;
  text-align: -internal-center;
}
</style>
<table>
  <tr>
    <th id="header"></th>
  </tr>
</table>
<script>
test(function() {
  assert_false(CSS.supports('text-align', '-internal-center'));
}, '"text-align" property does not support value "-internal-center"');
test(function() {
  assert_equals(getComputedStyle(header).textAlign, 'right');
}, '"text-align" property cannot be set to "-internal-center" by the author stylesheet');
</script>
